requirements.txt
pyspark==3.5.0
great_expectations==0.18.15
pyreadstat==1.2.6
pandas==2.1.4
streamlit==1.32.0
pyarrow==15.0.0
pyyaml==6.0.1
uuid
plotly==5.20.0
